refactor(agent): extract extract_content_or_reasoning into auxiliary_text (auxiliary_client.py R5-S1, epic #78647)
